Initial release of debian packaging files
This commit adds a debian folder that enables you to build a debian package for
dislocker. First make sure that you have the necessary software installed on
your system to build packages:

  sudo apt-get install build-essential debhelper devscripts quilt

Then you have to satisfy the build dependencies of dislocker:

  sudo apt-get install libfuse-dev libpolarssl-dev

Now to build the debian package go to the root folder of dislocker and execute:

  debuild -us -uc

When the build is done you might want to remove the build files with:

  fakeroot debian/rules clean

If nothing went wrong then both debian packages can be found at the parent
folder (../) and might be installed with:

  sudo dpkg -i libdislocker0_0.4.1-1_amd64.deb dislocker_0.4.1-1_amd64.deb

This has been tested with Debian Wheezy and Debian Jessie at 2015-01-26 for
dislocker of commit 646dfea22b82e396873fb62ee0896b342aabe015.